Improve search for closest objects in tl_detector.py
Previously, I was looking for nearest stopline and traffic light in
front of the car. However, such behaviour could be non-stable and
produce incorrect results when the car crossed the stop line, but had
RED traffic light corresponding to it in front.

In the current implementation, I first look for the nearest front
traffic light (to ensure it is in sight of the camera), then I search
for the stopline closest to the traffic light, then for waypoint closest
to the stopline. Such approach looks correct.

Signed-off-by: Dmitry Gavrilenko <dmitry.gavrilenko@icloud.com>