Apply minor improvements of tl classifier logic
These changes improve detection rate and response rate on simulator and
real bag-files. Some are designed specifically for the simulator.

- Increase MIN_SCORE_VALUE threshold to filter out false positive RED
traffic lights

- Resize input frame to 300x300 image (this is what neural net actually
does) to improve performance of further image operations

- Compare previous frame with existing frame content. If images are the
same, return previous result. This is a trick for simulator content to
minimize pipeline latency when the car stays in static mode, waiting for
GREEN light, which pops up for a few seconds

- Do not detect lights if they are outside of potential camera frustum
horizontal field of view, which is now hard-coded to 90 degrees

- Return RED_COUNT_THRESHOLD back to original value 1 to improve
response to RED traffic lights. There are less false positives for RED
traffic lights, therefore, no need to wait for 3 items in a row.

Signed-off-by: Dmitry Gavrilenko <dmitry.gavrilenko@icloud.com>